feat: Point 객체 생성 구현
